Consolidate all validators in one module
