+++ packages/taquito-michel-codec/src/taquito-michel-codec.ts
@@ -3,7 +3,5 @@ export * from "./micheline-parser";
 export * from "./micheline-emitter";
 export * from "./michelson-validator";
 export * from "./michelson-types";
-export * from "./michelson-typecheck-type";
-export * from "./michelson-typecheck-data";
-export * from "./michelson-typecheck-code";
+export * from "./michelson-typecheck";
 export { MacroError } from "./macros";
